fix: update avatar delete order
