+++ apps/journeys-admin/src/components/Editor/ControlPanel/Attributes/blocks/Footer/HostSidePanel/HostLocationFieldForm/HostLocationFieldForm.spec.tsx
@@ -71,6 +71,7 @@ describe('HostLocationFieldForm', () => {
 
     const { getByRole } = render(
       <MockedProvider
+        cache={cache}
         mocks={[
           {
             request: {
